Ajout du début de vérification des permissions dans les controllers de l'API. Début d'implémentation de l'authentication coté webapp à l'API

+++ bibliotheque-api/bibliotheque-api-service/src/main/java/com/bibliotheque/api/service/api/LivresApiController.java
@@ -1,6 +1,8 @@
 package com.bibliotheque.api.service.api;
 
 import com.bibliotheque.api.business.LivreManagement;
+import com.bibliotheque.api.business.UtilisateurManagement;
+import com.bibliotheque.api.model.Utilisateur;
 import com.bibliotheque.api.service.model.Livre;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.swagger.annotations.*;
@@ -9,6 +11,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -17,7 +21,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import javax.validation.Valid;
 import javax.servlet.http.HttpServletRequest;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -36,6 +39,9 @@ public class LivresApiController implements LivresApi {
     @Autowired
     private LivreManagement livreManagement;
 
+    @Autowired
+    private UtilisateurManagement utilisateurManagement;
+
     @org.springframework.beans.factory.annotation.Autowired
     public LivresApiController(ObjectMapper objectMapper, HttpServletRequest request) {
         this.objectMapper = objectMapper;
@@ -45,15 +51,26 @@ public class LivresApiController implements LivresApi {
     @Override
     public ResponseEntity<Void> addLivre(@ApiParam(value = "Un objet Livre doit être envoyé pour être ajouté", required = true) @Valid @RequestBody Livre body) {
         String accept = request.getHeader("Accept");
-        livreManagement.save(convertLivreApiToLivre(body));
-        return new ResponseEntity<Void>(HttpStatus.OK);
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (principal instanceof UserDetails) {
+            if (utilisateurManagement.findUtilisateurByMail(((UserDetails) principal).getUsername()).isBibliothecaire()) {
+                livreManagement.save(convertLivreApiToLivre(body));
+                return new ResponseEntity<Void>(HttpStatus.OK);
+            }
+        }
+        return new ResponseEntity<Void>(HttpStatus.UNAUTHORIZED);
     }
 
     public ResponseEntity<Void> deleteLivre(@ApiParam(value = "ID du livre à supprimer", required = true) @PathVariable("livreId") Long livreId, @ApiParam(value = "") @RequestHeader(value = "api_key", required = false) String apiKey) {
         String accept = request.getHeader("Accept");
-        if (livreManagement.findById(livreId).isPresent()) {
-            livreManagement.deleteById(livreId);
-            return new ResponseEntity<Void>(HttpStatus.OK);
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (principal instanceof UserDetails) {
+            if (utilisateurManagement.findUtilisateurByMail(((UserDetails) principal).getUsername()).isBibliothecaire()) {
+                if (livreManagement.findById(livreId).isPresent()) {
+                    livreManagement.deleteById(livreId);
+                    return new ResponseEntity<Void>(HttpStatus.OK);
+                }
+            }
         }
         return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
     }
@@ -79,13 +96,19 @@ public class LivresApiController implements LivresApi {
 
     public ResponseEntity<Void> updateLivre(@ApiParam(value = "ID du livre qui doit être mis à jour", required = true) @PathVariable("livreId") Long livreId, @ApiParam(value = "Mettre à jour le nom du livre") @RequestParam(value = "name", required = false) String name, @ApiParam(value = "Mettre à jour l'auteur d'un livre") @RequestParam(value = "auteur", required = false) String auteur, @ApiParam(value = "Mettre à jour la quantité d'un livre") @RequestParam(value = "quantite", required = false) Integer quantite) {
         String accept = request.getHeader("Accept");
-        Optional<com.bibliotheque.api.model.Livre> livre = livreManagement.findById(livreId);
-        if (livre.isPresent()) {
-            livre.get().setNom(name);
-            livre.get().setAuteur(auteur);
-            livre.get().setQuantite(quantite);
-            livreManagement.save(livre.get());
-            return new ResponseEntity<Void>(HttpStatus.OK);
+
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (principal instanceof UserDetails) {
+            if (utilisateurManagement.findUtilisateurByMail(((UserDetails) principal).getUsername()).isBibliothecaire()) {
+                Optional<com.bibliotheque.api.model.Livre> livre = livreManagement.findById(livreId);
+                if (livre.isPresent()) {
+                    livre.get().setNom(name);
+                    livre.get().setAuteur(auteur);
+                    livre.get().setQuantite(quantite);
+                    livreManagement.save(livre.get());
+                    return new ResponseEntity<Void>(HttpStatus.OK);
+                }
+            }
         }
         return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
     }
